requirements.txt
numpy==1.22.3
opencv-contrib-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
